+++ resources/js/components/form-loader/FormInput.vue
@@ -71,144 +71,135 @@ export default {
     dates: [],
     field_data: {}
   }),
-  created() {
+  created(){
     this.field_data = this.form_field;
   },
-  computed: {
+  computed:{
     additional_form: function() {
       return this.additional_form_data;
     }
   },
-  methods: {
-    display_col() {
+  methods:{
+    display_col(){
       let result = {};
-      if (
-        !this.is_undefined(this.field_data.type) &&
-        this.field_data.type == "hidden"
-      ) {
+      if (!this.is_undefined(this.field_data.type) && this.field_data.type == "hidden") 
+      {
         result["class"] = "hidden-input";
       }  
       return result;
     },
-    prepare_props(errors) {
+    prepare_props(errors){
       let result = {};
       result["error-messages"] = errors;
       result["label"] = this.prepare_label();
       result["outlined"] = true;
       result["dense"] = true;
 
-      if (
-        !this.is_undefined(this.field_data.help) &&
-        this.field_data.help !== ""
-      ) {
-        //result['prepend-inner-icon'] = this.field_data.value;
+      if (!this.is_undefined(this.field_data.help) && this.field_data.help !== "") 
+      {
+        result['prepend-inner-icon'] = this.field_data.value;
       }
 
-      if (
-        !this.is_undefined(this.field_data.type) &&
-        this.field_data.type == "time"
-      ) {
+      if (!this.is_undefined(this.field_data.type) && this.field_data.type == "time") 
+      {
         result["populate"] = this.field_data.value;
         result["ampm_in_title"] = true;
         result["no_title"] = false;
         result["class"] = "mt-2";
       }
 
-      if (
-        !this.is_undefined(this.field_data.type) &&
-        this.field_data.type == "file-input"
-      ) {
+      if (!this.is_undefined(this.field_data.type) && this.field_data.type == "file-input") 
+      {
         result["show-size"] = true;
         result["accept"] = this.field_data.accept;
       }
 
-      console.log( 'Data Type:' + this.field_data.type );
-      if (
-        !this.is_undefined(this.field_data.type) &&
-        (this.field_data.type == "select" ||
-          this.field_data.type == "customer-select" ||
-          this.field_data.type == "facility-select" ||
-          this.field_data.type == "zone-select" ||
-          this.field_data.type == "service-select")
-      ) {
-        result["selected"] =
-          !this.is_undefined(this.field_data.integer) &&
-          this.field_data.integer
-            ? parseInt(this.field_data.value)
-            : this.field_data.value;
-        result["any_field"] = !this.is_undefined(
-          this.field_data.any_field
-        )
-          ? this.field_data.any_field
-          : false;
-        result["items"] = this.field_data.items;
+      if (!this.is_undefined(this.field_data.type) && (this.field_data.type == "select")) 
+      {
+        result["selected"] = !this.is_undefined(this.field_data.integer) && this.field_data.integer
+          ? parseInt(this.field_data.value)
+          : this.field_data.value;
+
+        let select_items = this.field_data.items;
+        if(!this.is_undefined(this.field_data.any_field) && this.field_data.any_field)
+        {
+          let anyfield = {};
+          anyfield[this.field_data.item_value] = 'any';
+          anyfield[this.field_data.item_text] = 'Any';
+          //select_items.unshift( anyfield );
+        }
+        result["items"] = select_items;
         result["item-text"] = this.field_data.item_text;
         result["item-value"] = this.field_data.item_value;
         result["data-vv-name"] = "select";
         result["chips"] = true;
       }
-      if (
-        !this.is_undefined(this.field_data.type) &&
-        this.field_data.type == "hidden"
-      ) {
+      if ( !this.is_undefined(this.field_data.type) && this.field_data.type == "hidden" )
+      {
         result["class"] = "hidden-input";
       }
 
-      if (
-        !this.is_undefined(this.field_data.component_type) &&
-        this.field_data.component_type == "password"
-      ) {
+      if (!this.is_undefined(this.field_data.component_type) && this.field_data.component_type == "password") 
+      {
         result["type"] = this.field_data.component_type;
       }
 
-      if (
-        !this.is_undefined(this.field_data.component) &&
-        this.field_data.component == "date-picker"
-      ) {
+      if (!this.is_undefined(this.field_data.component) && this.field_data.component == "date-picker") 
+      {
         result["close-on-content-click"] = false;
         result["controls"] = false;
         if (!this.is_undefined(this.field_data.close_on_content_click))
-          result[
-            "close-on-content-click"
-          ] = this.field_data.close_on_content_click;
-
+        {
+          result["close-on-content-click"] = this.field_data.close_on_content_click;
+        }
         if (!this.is_undefined(this.field_data.controls))
+        {
           result["controls"] = this.field_data.controls;
+        }
       }
 
-      if (!this.is_undefined(this.field_data.step)) {
+      if (!this.is_undefined(this.field_data.step)) 
+      {
         result["step"] = this.field_data.step;
       }
 
-      if (!this.is_undefined(this.field_data.component_type)) {
+      if (!this.is_undefined(this.field_data.component_type) && this.field_data.component_type !== '') 
+      {
         result["type"] = this.field_data.component_type;
       }
 
-      if (!this.is_undefined(this.field_data.min)) {
+      if (!this.is_undefined(this.field_data.min)) 
+      {
         result["min"] = this.field_data.min;
       }
 
-      if (!this.is_undefined(this.field_data.max)) {
+      if (!this.is_undefined(this.field_data.max)) 
+      {
         result["max"] = this.field_data.max;
       }
 
-      if (!this.is_undefined(this.field_data.placeholder)) {
+      if (!this.is_undefined(this.field_data.placeholder)) 
+      {
         result["placeholder"] = this.field_data.placeholder;
       }
 
-      if (!this.is_undefined(this.field_data.readonly)) {
+      if (!this.is_undefined(this.field_data.readonly)) 
+      {
         result["readonly"] = this.field_data.readonly;
       }
 
-      if (!this.is_undefined(this.field_data.multiple)) {
+      if (!this.is_undefined(this.field_data.multiple)) 
+      {
         result["multiple"] = this.field_data.multiple;
       }
 
-      if (!this.is_undefined(this.additional_form_data)) {
+      if (!this.is_undefined(this.additional_form_data)) 
+      {
         result["additional_form_data"] = this.additional_form;
       }
 
-      if (this.field_data.component == "h2") {
+      if (this.field_data.component == "h2") 
+      {
         result = { "v-html": this.field_data.value };
       }
 
@@ -218,20 +209,21 @@ export default {
       let label = !this.is_undefined(this.field_data.label)
         ? this.field_data.label
         : this.field_data.name;
-      if (
-        !this.is_undefined(this.field_data.required) &&
-        this.field_data.required
-      ) {
+
+      if (!this.is_undefined(this.field_data.required) && this.field_data.required) 
+      {
         label += " ( * required )";
-      } else {
+
+      } 
+      else if (!this.is_undefined(this.field_data.readonly) && this.field_data.readonly) 
+      {
+        label += " ( readonly )";
+      } 
+      else 
+      {
         label += " ( optional )";
       }
-      if (
-        !this.is_undefined(this.field_data.readonly) &&
-        this.field_data.readonly
-      ) {
-        label = this.field_data.label + " ( readonly )";
-      }
+
       return label;
     }
   }
@@ -239,14 +231,17 @@ export default {
 </script>
 
 <style>
-.v-select__selections {
+.v-select__selections 
+{
   margin-top: 3px !important;
   margin-bottom: 2px !important;
 }
-.v-select.v-input--dense .v-chip {
+.v-select.v-input--dense .v-chip 
+{
   margin-bottom: 2px !important;
 }
-.hidden-input{
+.hidden-input
+{
   display: none;
 }
 </style>