+++ resources/js/components/form-loader/FormMixins.js
@@ -3,38 +3,40 @@ import axios from 'axios';
 export const FormMixin = {
     methods: {
         is_object_empty: function(o) {
-        for (var p in o) {
-            if (Object.prototype.hasOwnProperty.call(o,p)) {
-            return false; 
+        for (var p in o) 
+        {
+            if (Object.prototype.hasOwnProperty.call(o,p)) 
+            {
+                return false; 
             } 
         }
         return true;
         },
         is_object: function(o) { 
-        return typeof o === "object" 
+            return typeof o === "object" 
         },
         is_array: function(o) { 
-        return Array.isArray(o) 
+            return Array.isArray(o) 
         },
         is_file: function(o) { 
-        return('File' in window && o instanceof File) ? true : false 
+            return('File' in window && o instanceof File) ? true : false 
         },
         is_undefined: function(o) { 
-        return typeof(o) === 'undefined' 
+            return typeof(o) === 'undefined' 
         },
         snake_to_camel: function(s) {
-        return s.toLowerCase().replace(/[-_][a-z0-9]/g, (group) => group.slice(-1).toUpperCase());
+            return s.toLowerCase().replace(/[-_][a-z0-9]/g, (group) => group.slice(-1).toUpperCase());
         },
         request: function(type, url, data = null, expecting_results = false , showalert = true , parameters = {} , loader = true) {
-        var _this = this;
-        let default_parameters = { headers: { 'Content-Type': 'application/json' } };
-        let send_parameters = { headers: { ...default_parameters, ...parameters } };
-        let return_reponse = {};
-        return axios[type](url, data, send_parameters)
+            var _this = this;
+            let default_parameters = { headers: { 'Content-Type': 'application/json' } };
+            let send_parameters = { headers: { ...default_parameters, ...parameters } };
+            let return_reponse = {};
+            return axios[type](url, data, send_parameters)
             .then(response => {
-                var request_response = response.data;
+                let request_response = response.data;
                 if(this.is_object(request_response)) {
-                    if(! request_response.success) {
+                    if(! request_response.result) {
                         if(! this.is_object(request_response.data)) {
                             if(showalert)
                                 return_reponse.alert = { text: request_response.data , colour: 'error' };