+++ src/Base/EasyForm.php
@@ -2,15 +2,12 @@
 namespace PlusTimeIT\EasyForms\Base;
 
 use PlusTimeIT\EasyForms\Interfaces\FormInterface;
-use PlusTimeIT\EasyForms\Traits\FormTrait;
 
-class EasyForm implements FormInterface
+abstract class EasyForm implements FormInterface
 {
     protected $fields = [];
 
     protected $name = '';
 
     protected $title = '';
-
-    use FormTrait;
 }