+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All version change notes to be placed here.
 
 ## Version Updates
 
+### 0.0.5
+---
+-  Fixing FormInput and FormLoader vue files
+-  Creation of AxiosResponse Element, Interface and Trait
+
 ### 0.0.4
 ---
 -  Fixed bug: FileInputField not displaying file dialog box #5